test: 重构 AuthorizationViewTestBase 单元测试
