+++ test/UnitTest/Components/AuthorizeViewTest.cs
@@ -8,7 +8,7 @@ using UnitTest.Pages;
 
 namespace UnitTest.Components;
 
-public class AuthorizeViewTest : AuthorizateViewTestBase
+public class AuthorizeViewTest : AuthorizationViewTestBase
 {
     [Fact]
     public void NotAuthorized_Ok()
