+++ test/UnitTest/Components/LayoutTest.cs
@@ -385,7 +385,7 @@ public class LayoutTest : BootstrapBlazorTestBase
     private static RenderFragment CreateSide(string? content = "Side") => builder => builder.AddContent(0, content);
 }
 
-public class LayoutAuthorizationTest : AuthorizateViewTestBase
+public class LayoutAuthorizationTest : AuthorizationViewTestBase
 {
     [Fact]
     public void Authorized_Ok()
